requirements.txt
python-dotenv==1.0.1
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.16
streamlit==1.36.0
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
pypdf==4.2.0